Abstractions

Functions

DynOptInterface.AbstractDynamicFunction — Type

AbstractDynamicFunction <: MOI.AbstractScalarFunction

Abstract supertype for dynamic functions.

DynOptInterface.AbstractAlgebraicFunction — Type

AbstractAlgebraicFunction <: AbstractDynamicFunction

Abstract supertype for algebraic functions. That is, expressions that may contain $y(t)$ or $t$, but not $\dot{y}(t)$.

DynOptInterface.AbstractDifferentialFunction — Type

AbstractDifferentialFunction <: AbstractDynamicFunction

Abstract supertype for differential functions. That is, expressions that may contain $\dot{y}(t)$, $y(t)$, or $t$.

DynOptInterface.AbstractBoundaryFunction — Type

AbstractBoundaryFunction <: AbstractDynamicFunction

Abstract supertype for functions evaluated at domain boundaries. That is, expressions that may contain $y(t^0)$, $y(t^f)$, $t^0$, or $t^f$.

Attributes

DynOptInterface.AbstractDomainAttribute — Type

AbstractDomainAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of domains in the model.

DynOptInterface.AbstractDynamicVariableAttribute — Type

AbstractDynamicVariableAttribute

Abstract supertype for attributs objects that can be used to set or get attributes (properties) of dynamic variables in the model.