Algebraic Functions

Types

DynOptInterface.LinearAlgebraicTerm — Type

LinearAlgebraicTerm{T}

\[t_i \mapsto c_j y_j(t_i)\]

DynOptInterface.LinearAlgebraicFunction — Type

LinearAlgebraicFunction{T} <: AbstractAlgebraicFunction

\[t_i \mapsto c^\top y(t_i)\]

DynOptInterface.SquaredAlgebraicTerm — Type

SquaredAlgebraicTerm{T}

\[t_i \mapsto y_j(t_i) c_{jk} y_k(t_i)\]

DynOptInterface.SquaredAlgebraicFunction — Type

SquaredAlgebraicFunction{T}

\[t_i \mapsto y(t_i)^\top C y(t_i)\]

DynOptInterface.NonlinearAlgebraicFunction — Type

NonlinearAlgebraicFunction <: AbstractAlgebraicFunction

\[t_i \mapsto f_a(y(t_i), t_i, x)\]

Similar to MathOptInterface.ScalarNonlinearFunction, an expression tree is used to represent nonlinear functions.

struct NonlinearAlgebraicFunction <: AbstractAlgebraicFunction
    head::Symbol
    args::Vector{Any}
    t_i::DomainIndex
...
end

head

The symbol head must be an operator that is supported by the model. The model attribute MathOptInterface.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, a MathOptInterface.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. Each element in args can be one of the following:

A constant value of type T<:Real

A MathOptInterface.VariableIndex $x_k$

A MathOptInterface.ScalarAffineFunction $a^\top x + b$

A MathOptInterface.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

A MathOptInterface.ScalarNonlinearFunction $f(x)$

A DomainIndex $t_i$

A DynamicVariableIndex $y_j(t_i)$

A LinearAlgebraicFunction $c^\top y(t_i)$

A SquaredAlgebraicFunction $y(t_i)^\top C y(t_i)$

Another NonlinearAlgebraicFunction

t_i

As expressions are composed, dynamic functions must have the same DomainIndex t_i.