Attributes

Variable Attributes

The following table defines which DynOptInterface objects are compatible with sub-types of MOI.AbstractVariableAttributes.

Attribute	$t$	$t^0$	$t^f$	$y(\cdot)$	$y(t^0)$	$y(t^f)$

MOI.VariableName	✅	❌	❌	✅	❌	❌

MOI.VariablePrimal	❌	✅	✅	✅	✅	✅

MOI.VariablePrimalStart	❌	✅	✅	✅	✅	✅

MathOptInterface.supports — Function

MOI.supports(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    ::Type{PhaseIndex},
)::Bool

Return a Bool indicating whether model supports the attribute attr for PhaseIndexs $t$.

    MOI.supports(
        model::MOI.ModelLike,
        attr::MOI.AbstractVariableAttribute,
        ::Type{Initial{PhaseIndex}},
    )::Bool

Return a Bool indicating whether model supports the attribute attr for Initial{PhaseIndex}s $t^0$.

    MOI.supports(
        model::MOI.ModelLike,
        attr::MOI.AbstractVariableAttribute,
        ::Type{Final{PhaseIndex}},
    )::Bool

Return a Bool indicating whether model supports the attribute attr for Final{PhaseIndex}s $t^f$.

    MOI.supports(
        model::MOI.ModelLike,
        attr::MOI.AbstractVariableAttribute,
        ::Type{DynamicVariableIndex},
    )::Bool

Return a Bool indicating whether model supports the attribute attr for DynamicVariableIndexs $y(\cdot)$.

    MOI.supports(
        model::MOI.ModelLike,
        attr::MOI.AbstractVariableAttribute,
        ::Type{Initial{DynamicVariableIndex}},
    )::Bool

Return a Bool indicating whether model supports the attribute attr for Initial{DynamicVariableIndex}s $y(t^0)$.

    MOI.supports(
        model::MOI.ModelLike,
        attr::MOI.AbstractVariableAttribute,
        ::Type{Final{DynamicVariableIndex}},
    )::Bool

Return a Bool indicating whether model supports the attribute attr for Final{DynamicVariableIndex}s $y(t^f)$.

MathOptInterface.set — Function

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i::PhaseIndex,
    value,
)

Assign value to the attribute attr of phase t_i in model model.

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j::DynamicVariableIndex,
    value,
)

Assign value to the attribute attr of dynamic variable y_j in model model.

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i_0::Initial{PhaseIndex},
    value,
)

Assign value to the attribute attr of initial phase t_i_0 in model model.

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i_f::Final{PhaseIndex},
    value,
)

Assign value to the attribute attr of final phase t_i_f in model model.

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j_0::Initial{DynamicVariableIndex},
    value,
)

Assign value to the attribute attr of initial phase y_j_0 in model model.

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j_f::Final{DynamicVariableIndex},
    value,
)

Assign value to the attribute attr of final phase y_j_f in model model.

For all of the above methods, an MOI.UnsupportedAttribute error is thrown if model does not support the attribute attr, and a MOI.SetAttributeNotAllowed error is thrown if it supports the attribute attr but it cannot be set.

MathOptInterface.get — Function

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i::PhaseIndex,
)

Return the value of the attribute attr set to phase t_i in model model.

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j::DynamicVariableIndex,
)

Return the value of the attribute attr set to dynamic variable y_j in model model.

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i_0::Initial{PhaseIndex},
)

Return the value of the attribute attr set to the initial phase t_i_0 in model model.

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    t_i_f::Final{PhaseIndex},
)

Return the value of the attribute attr set to the final phase t_i_f in model model.

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j_0::Initial{DynamicVariableIndex},
)

Return the value of the attribute attr set to the initial dynamic variable y_j_0 in model model.

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    y_j_f::Final{DynamicVariableIndex},
)

Return the value of the attribute attr set to the final dynamic variable y_j_f in model model.

For all of the above methods, if the attribute attr is not supported by model then an error should be thrown. If the attribute is supported but has not been set, nothing is returned.