Boundary Functions

DynOptInterface.AbstractBoundaryFunction — Type

AbstractBoundaryFunction <: MOI.AbstractScalarFunction

Abstract supertype for dynamic functions evaluated at phase boundaries. That is, expressions that may contain $t^0$, $t^f$, $y(t^0)$, or $y(t^f)$.

DynOptInterface.Initial — Type

Initial{AF<:AbstractDynamicFunction} <: AbstractBoundaryFunction

\[b^0(y(t_i^0), t_i^0, x)\]

Represents the evaluation of an AbstractDynamicFunction at the initial point of its phase. Common cases are:

$t_i^0$ Initial{PhaseIndex}

$y_j(t_i^0)$ Initial{DynamicVariableIndex}

DynOptInterface.Final — Type

Final{AF<:AbstractDynamicFunction} <: AbstractBoundaryFunction

\[b^f(y(t_i^f), t_i^f, x)\]

Represents the evaluation of an AbstractDynamicFunction at the final point of its phase. Common cases are:

$t_i^f$ Final{PhaseIndex}

$y_j(t_i^f)$ Final{DynamicVariableIndex}

DynOptInterface.NonlinearBoundaryFunction — Type

struct NonlinearBoundaryFunction <: AbstractBoundaryFunction
    head::Symbol
    args::Vector{Any}
...
end

\[f_b(y(t^0), y(t^f), t^0, t^f, x)\]

Similar to MOI.ScalarNonlinearFunction, nonlinear dynamic functions are represented by an expression tree.

head

The symbol head must be an operator that is supported by the model. The model attribute MOI.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, a MOI.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. The possible arguments that may be included are:

A constant value of type T<:Real

A MOI.VariableIndex $x_k$

A MOI.ScalarAffineFunction $a^\top x + b$

A MOI.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

A MOI.ScalarNonlinearFunction $f(x)$

A Initial of PhaseIndex $t_i^0$

A Final of PhaseIndex $t_i^f$

A Initial of DynamicVariableIndex $y(t_i^0)$

A Final of DynamicVariableIndex $y(t_i^f)$

Another NonlinearBoundaryFunction

Additionally, the optimizer must indicate support of argument types through the supports_objective_argument and supports_constraint_argument functions.