Domains

DynOptInterface.AbstractDomain — Type

AbstractDomain

Abstract supertype for domains.

DynOptInterface.DomainIndex — Type

DomainIndex{D}

A type-safe wrapper for Int64 for use in referencing domains. The parameter D is the type of the domain.

DynOptInterface.supports_domain — Function

supports_domain(
    model::MOI.ModelLike,
    ::Type{D},
) where {D<:AbstractDomain}

Return a Bool indicating whether model supports domains of type D.

DynOptInterface.UnsupportedDomain — Type

struct UnsupportedDomain{D<:AbstractDomain} <: MOI.UnsupportedError
    message::String
end

An error indicating that domains of type D are not supported by the model, that is, that supports_domain returns false.

DynOptInterface.AddDomainNotAllowed — Type

struct AddDomainNotAllowed{D<:AbstractDomain} <: MOI.NotAllowedError
    message::String
end

An error indicating that domains of type D are supported but cannot be added to the current state of the model.

DynOptInterface.add_domain — Function

add_domain(
    model::MOI.ModelLike,
    domain::AbstractDomain
)

Add domain to the model. An UnsupportedDomain error is thrown if model does not support. Otherwise a DomainIndex is returned.

DynOptInterface.Interval — Type

struct Interval{T0, TF} <: AbstractDomain
    t_0::T0
    t_f::TF
end

A one-dimensional domain where t_0 and t_f may be real numbers or variable indices.