Dynamic Functions

DynOptInterface.LinearDynamicTerm — Type

LinearDynamicTerm{T}

\[t_i \mapsto c_j y_j(t_i)\]

DynOptInterface.LinearDynamicFunction — Type

LinearDynamicFunction{T} <: AbstractDynamicFunction

\[t_i \mapsto c^\top y(t_i)\]

DynOptInterface.SquaredDynamicTerm — Type

SquaredDynamicTerm{T}

\[t_i \mapsto y_j(t_i) c_{jk} y_k(t_i)\]

DynOptInterface.SquaredDynamicFunction — Type

SquaredDynamicFunction{T} <: AbstractDynamicFunction

\[t_i \mapsto y(t_i)^\top C y(t_i)\]

DynOptInterface.NonlinearDynamicFunction — Type

struct NonlinearDynamicFunction <: AbstractDynamicFunction
    head::Symbol
    args::Vector{Any}
...
end

\[t_i \mapsto f_d(\dot{y}(t_i), y(t_i), t_i, x)\]

Similar to MOI.ScalarNonlinearFunction, nonlinear dynamic functions are represented by an expression tree.

head

The symbol head must be an operator that is supported by the model. The model attribute MOI.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, a MOI.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. The possible arguments that may be included are:

A constant value of type T<:Real

A MOI.VariableIndex $x_k$

A MOI.ScalarAffineFunction $a^\top x + b$

A MOI.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

A MOI.ScalarNonlinearFunction $f(x)$

A PhaseIndex $t_i$

A DynamicVariableIndex $y_j(t_i)$

A LinearDynamicFunction $c^\top y(t_i)$

A SquaredDynamicFunction $y(t_i)^\top C y(t_i)$

Another NonlinearDynamicFunction

Additionally, the optimizer must indicate support of argument types through the supports_objective_argument and supports_constraint_argument functions.

DynOptInterface.ExplicitDifferentialFunction — Type

ExplicitDifferentialFunction{AF<:AbstractDynamicFunction} <: AbstractDynamicFunction

\[t_i \mapsto \dot{y}(t_i) - a(y(t_i), t_i, x)\]