Dynamic Variables

DynOptInterface.DynamicVariableIndex — Type

DynamicVariableIndex(value::Int64, phase::DynamicVariableIndex)

An object for use in referencing dynamic variables in a model.

It is a sub-type of AbstractDynamicFunction. It represents a dynamic variable $t_i \mapsto y_j(t_i)$ defined in the domain $[t_i^0, t_i^f]$. The Int64 index value is stored in the value field. To allow for deletion, indices need not be consecutive.

DynOptInterface.supports_dynamic_variable — Function

supports_dynamic_variable(model::MOI.ModelLike)::Bool

Return a Bool indicating whether model supports dynamic variables.

DynOptInterface.UnsupportedDynamicVariable — Type

UnsupportedDynamicVariable(message::String)

An error indicating that dynamic variables are not supported by the model.

That is, that supports_dynamic_variable returns false. The message String is stored in the message field.

DynOptInterface.AddDynamicVariableNotAllowed — Type

AddDynamicVariableNotAllowed(message::String)

An error indicating that dynamic variables cannot be added to the model in its current state.

The String error message is stored in the message field.

DynOptInterface.add_dynamic_variable — Function

add_dynamic_variable(model::MOI.ModelLike, phase::PhaseIndex)::DynamicVariableIndex

Add a dynamic variable to model, returning a DynamicVariableIndex.

An AddDynamicVariableNotAllowed is thrown if a dynamic variable cannot be added to model in its current state.

MathOptInterface.is_valid — Method

MOI.is_valid(model::MOI.ModelLike, index::DynamicVariableIndex)::Bool

Return a Bool indicating whether index refers to a valid DynamicVariableIndex in model.

DynOptInterface.InvalidDynamicVariableIndex — Type

InvalidDynamicVariableIndex(index::DynamicVariableIndex)

An error indicating that the dynamic variable index is not valid.

Attributes

The DynamicVariableIndex object is compatible with the following attributes:

MOI.VariableName with value types String

MOI.VariablePrimal with value types AbstractDynamicSolution

MOI.VariablePrimalStart with value types AbstractDynamicSolution

MathOptInterface.supports — Method

MOI.supports(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    ::Type{DynamicVariableIndex},
)::Bool

Return a Bool indicating whether model supports the attribute attr for DynamicVariableIndexs.

MathOptInterface.set — Method

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    index::DynamicVariableIndex,
    value,
)

Assign value to the attribute attr of dynamic variable index in model model.

An MOI.UnsupportedAttribute error is thrown if model does not support the attribute attr, and a MOI.SetAttributeNotAllowed error is thrown if it supports the attribute attr but it cannot be set.

MathOptInterface.get — Method

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    index::DynamicVariableIndex,
)

Return the value of the attribute attr set to dynamic variable index in model model.

If the attribute attr is not supported by model then an error should be thrown. If the attribute is supported but has not been set, nothing is returned.