Expressions

Linear

DynOptInterface.LinearDynamicTerm — Type

LinearDynamicTerm{T}(coefficient::T, dyn_var::DynamicVariableIndex) where {T}

Represents the term $c_j y_j$, where $c_j$ is a coefficient and $y_j$ is a DynamicVariableIndex.

The coefficient is stored in the coefficient field and the DynamicVariableIndex is stored in the dyn_var field.

DynOptInterface.LinearDynamicFunction — Type

LinearDynamicFunction{T}(terms::Vector{LinearDynamicTerm{T}}) where {T}

Represents the function $t_i \mapsto c^\top y(t_i)$, which is a sum of LinearDynamicTerms.

A sub-type of AbstractDynamicFunction. All dynamic variables must be defined on the same phase, otherwise a MixedPhases error is thrown. The LinearDynamicTerms are stored in the terms field.

Squared

DynOptInterface.SquaredDynamicTerm — Type

SquaredDynamicTerm{T}(
    coefficient::T
    dyn_var_a::DynamicVariableIndex
    dyn_var_b::DynamicVariableIndex
) where {T}

Represents the term $c_{ab} y_a y_b$, where $c_{ab}$ is a coefficient and $y_a, y_b$ are DynamicVariableIndexs.

The coefficient is stored in the coefficient field. The dynamic variables are stored in the dyn_var_a and dyn_var_b fields and must be defined on the same phase, otherwise a MixedPhases error is thrown.

DynOptInterface.SquaredDynamicFunction — Type

SquaredDynamicFunction{T}(terms::Vector{SquaredDynamicTerm{T}}) where {T}

Represents the function $t_i \mapsto y(t_i)^\top C y(t_i)$, which is a sum of SquaredDynamicTerms.

A sub-type of AbstractDynamicFunction. All dynamic variables must be defined on the same phase, otherwise a MixedPhases error is thrown. The SquaredDynamicTerms are stored in the terms field.

Nonlinear

DynOptInterface.NonlinearDynamicFunction — Type

NonlinearDynamicFunction(head::Symbol, args::Vector{Any}, phase::PhaseIndex)

Represents a general function $t_i \mapsto f_d(\dot{y}(t_i), y(t_i), t_i, x)$.

A sub-type of AbstractDynamicFunction. All dynamic variables must be defined on the same phase, otherwise a MixedPhases error is thrown. Similar to MOI.ScalarNonlinearFunction, nonlinear dynamic functions are represented by expression trees, using the following fields:

head

The symbol head must be an operator that is supported by the model. The model attribute MOI.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, an MOI.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. The possible arguments that may be included are:

A constant value of type T<:Real

An MOI.VariableIndex $x_k$

An MOI.ScalarAffineFunction $a^\top x + b$

An MOI.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

An MOI.ScalarNonlinearFunction $f(x)$

A PhaseIndex $t_i$

A DynamicVariableIndex $y_j(t_i)$

A LinearDynamicFunction $c^\top y(t_i)$

A SquaredDynamicFunction $y(t_i)^\top C y(t_i)$

Another NonlinearDynamicFunctions

Additionally, the optimizer must indicate support of argument types through the supports_objective_argument and supports_constraint_argument functions.