Dynamic Variables

Type

DynOptInterface.DynamicVariableIndex — Type

DynamicVariableIndex <: AbstractDynamicFunction

\[t_i \mapsto y_j(t_i)\]

A type-safe wrapper for Int64 for use in referencing dynamic variables.

Errors

DynOptInterface.UnsupportedDynamicVariable — Type

UnsupportedDynamicVariable <: MOI.UnsupportedError

An error indicating that dynamic variables are not supported by the model, that is, that supports_dynamic_variable returns false.

DynOptInterface.AddDynamicVariableNotAllowed — Type

AddDynamicVariableNotAllowed <: MOI.NotAllowedError

An error indicating that dynamic variables cannot be added to the model in its current state.

Functions

DynOptInterface.supports_dynamic_variable — Function

supports_dynamic_variable(model::MOI.ModelLike)

Return a Bool indicating whether model supports dynamic variables.

DynOptInterface.add_dynamic_variable — Function

add_dynamic_variable(model::MOI.ModelLike)

Add a dynamic variable to model, returning a DynamicVariableIndex. An AddDynamicVariableNotAllowed is thrown if a dynamic variable cannot be added to model in its current state.

Attributes

DynOptInterface.DynamicVariableName — Type

DynamicVariableName <: AbstractDomainAttribute

A dynamic variable attribute for a String identifying a DynamicVariableIndex. If not set, it has a default value of "".

DynOptInterface.DynamicVariablePrimalStart — Type

DynamicVariablePrimalStart <: AbstractDynamicVariableAttribute

A dynamic variable attribute for setting a starting function for $y_j(t_i)$, which may help warm-start the optimizer. It is either a nothing (unset) or a single-argument function of T<:Real.

DynOptInterface.DynamicVariablePrimal — Type

DynamicVariablePrimal <: AbstractDynamicVariableAttribute

A dynamic variable attribute for setting or getting $y_j(t_i)$. It should be a single-argument function of T<:Real. It should have the same behaviour as the MathOptInterface.VariablePrimal attribute.