Nonlinear Support

DynOptInterface.supports_objective_argument — Function

supports_objective_argument(
    model::MOI.ModelLike,
    F::Type{<:MOI.AbstractScalarFunction},
    A::Type{<:MOI.AbstractScalarFunction},
)::Bool

Return a Bool indicating whether model supports objective functions of type F containing one or more arguments of type A.

DynOptInterface.UnsupportedObjectiveArgument — Type

UnsupportedObjectiveArgument{F,A}(message::String)

An error indicating that supports_objective_argument returns false.

The String error message is stored in the message field.

DynOptInterface.supports_constraint_argument — Function

supports_constraint_argument(
    model::MOI.ModelLike,
    F::Type{<:MOI.AbstractScalarFunction},
    S::Type{<:MOI.AbstractScalarSet},
    A::Type{<:MOI.AbstractScalarFunction},
)::Bool

Return a Bool indicating whether model supports F-in-S constraints, where the function contains one or more arguments of type A.

DynOptInterface.UnsupportedConstraintArgument — Type

UnsupportedConstraintArgument{F,S,A}(message::String)

An error indicating that supports_constraint_argument returns false.

The String error message is stored in the message field.