Phases

Types

DynOptInterface.AbstractDynamicFunction — Type

AbstractDynamicFunction <: MOI.AbstractScalarFunction

Abstract supertype for dynamic functions. That is, expressions that may contain $t$, $y(t)$, or $\dot{y}(t)$.

DynOptInterface.PhaseIndex — Type

struct PhaseIndex <: AbstractDynamicFunction
    value::Int64
end

\[t_i \in [t_i^0, t_i^f]\]

A type-safe wrapper for Int64 for use in referencing phases.

Errors

DynOptInterface.UnsupportedPhase — Type

UnsupportedPhase <: MOI.UnsupportedError

An error indicating that phases are not supported by the model, that is, that supports_phase returns false.

DynOptInterface.AddPhaseNotAllowed — Type

AddPhaseNotAllowed <: MOI.NotAllowedError

An error indicating that phases cannot be added to the model in its current state.

Functions

DynOptInterface.supports_phase — Function

supports_phase(model::MOI.ModelLike)::Bool

Returns a Bool indicating whether model supports phases.

DynOptInterface.add_phase — Function

add_phase(model::MOI.ModelLike)

Add a phase to model, returning a PhaseIndex. An AddPhaseNotAllowed error is thrown if a phase cannot be added to the model in its current state.

MathOptInterface.supports — Method

MOI.supports(
    model::MOI.ModelLike,
    attr::AbstractPhaseAttribute,
    ::Type{PhaseIndex},
)::Bool

Returns a Bool indicating whether model supports the phase attribute attr.

MathOptInterface.set — Method

MOI.set(
    model::MOI.ModelLike,
    attr::AbstractPhaseAttribute,
    t_i::PhaseIndex,
    value,
)

Assigns value to the attribute attr of phase t_i in model model. An MOI.UnsupportedAttribute error is thrown if model does not support the attribute attr, and a MOI.SetAttributeNotAllowed error is thrown if it supports the attribute attr but it cannot be set.

MathOptInterface.get — Method

MOI.get(model::MOI.ModelLike, attr::AbstractPhaseAttribute, t_i::PhaseIndex)

If the attribute attr is set for the phase t_i, return its value; return nothing otherwise. If the attribute attr is not supported by model, an error is thrown.

Attributes

DynOptInterface.AbstractPhaseAttribute — Type

AbstractPhaseAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of phases in the model.

DynOptInterface.PhaseName — Type

PhaseName <: AbstractPhaseAttribute

A phase attribute for a String identifying a PhaseIndex. If not set, it has a default value of "".

DynOptInterface.PhaseInitialPrimalStart — Type

PhaseInitialPrimalStart <: AbstractPhaseAttribute

A phase attribute for setting a starting value for $t_i^0$, which may help warm-start the optimizer. It is either a number or a nothing (unset).

DynOptInterface.PhaseFinalPrimalStart — Type

PhaseFinalPrimalStart <: AbstractPhaseAttribute

A phase attribute for setting a starting value for $t_i^f$, which may help warm-start the optimizer. It is either a number or a nothing (unset).

DynOptInterface.PhaseInitialPrimal — Type

PhaseInitialPrimal <: AbstractPhaseAttribute

A phase attribute for setting or getting $t_i^0$. It should have the same behaviour as the MOI.VariablePrimal attribute.

DynOptInterface.PhaseFinalPrimal — Type

PhaseFinalPrimal <: AbstractPhaseAttribute

A phase attribute for setting or getting $t_i^f$. It should have the same behaviour as the MOI.VariablePrimal attribute.