Boundary Functions

Abstraction

DynOptInterface.AbstractBoundaryFunction — Type

AbstractBoundaryFunction

Abstract supertype for dynamic functions evaluated at phase boundaries.

That is, expressions that may be or contain:

$t^0$ Initial{PhaseIndex}

$t^f$ Final{PhaseIndex}

$y(t^0)$ Initial{DynamicVariableIndex}

$y(t^f)$ Final{DynamicVariableIndex}

Initial & Final

DynOptInterface.Initial — Type

Initial{DF}(dyn_fun::DF) where {DF<:AbstractDynamicFunction}

Represents the evaluation of an AbstractDynamicFunction at the initial point of its phase.

A sub-type of AbstractBoundaryFunction. The dynamic function is stored in the dyn_fun field.

DynOptInterface.Final — Type

Final{DF}(dyn_fun::DF) where {DF<:AbstractDynamicFunction}

Represents the evaluation of an AbstractDynamicFunction at the final point of its phase.

A sub-type of AbstractBoundaryFunction. The dynamic function is stored in the dyn_fun field.

Linkage

DynOptInterface.Linkage — Type

Linkage{DF}(
    final::Final{DF},
    initial::Initial{DF},
) where {DF<:AbstractDynamicFunction}

Represents the expression $f_f(y(t^f), t^f) - f_0(y(t^0), t^0)$.

A sub-type of AbstractBoundaryFunction. The final function is stored in the final field and the initial function is stored in the initial field.

Nonlinear

DynOptInterface.NonlinearBoundaryFunction — Type

NonlinearBoundaryFunction(head::Symbol, args::Vector{Any})

Represents a general function $f_b(y(t^0), y(t^f), t^0, t^f, x)$.

A sub-type of AbstractBoundaryFunction. Similar to MOI.ScalarNonlinearFunction, nonlinear boundary functions are represented by expression trees, using the following fields:

head

The symbol head must be an operator that is supported by the model. The model attribute MOI.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, an MOI.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. The possible arguments that may be included are:

A constant value of type T<:Real

An MOI.VariableIndex $x_k$

An MOI.ScalarAffineFunction $a^\top x + b$

An MOI.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

An MOI.ScalarNonlinearFunction $f(x)$

An Initial of PhaseIndex $t_i^0$

A Final of PhaseIndex $t_i^f$

An Initial of DynamicVariableIndex $y(t_i^0)$

A Final of DynamicVariableIndex $y(t_i^f)$

Another NonlinearBoundaryFunctions

Additionally, the optimizer must indicate support of argument types through the supports_objective_argument and supports_constraint_argument functions.

Integrals

DynOptInterface.Integral — Type

Integral{DF}(
    dyn_fun::AbstractDynamicFunction,
) where {DF<:AbstractDynamicFunction}

Represents the integral $\int_{t_i^o}^{t_i^f} f_d(\dot{y}(t_i), y(t_i), t_i, x) \mathrm{d}t_i$.

A sub-type of AbstractBoundaryFunction. The integrand is stored in the dyn_fun field.

DynOptInterface.Bolza — Type

Bolza{BF,DF}(
    bou_fun::BF,
    integral::Integral{DF},
) where {BF<:AbstractBoundaryFunction,DF<:AbstractDynamicFunction}

\[f_b(y_0, y_f, t_0, t_f, x) + \int_{t_i^o}^{t_i^f} f_d(\dot{y}(t_i), y(t_i), t_i, x) \mathrm{d}t_i\]

Represents the sum of an AbstractBoundaryFunction with the integral of an AbstractDynamicFunction.

A sub-type of AbstractBoundaryFunction. The boundary function is stored in the bou_fun field and the integral is stored in the integral field.