Boundary Functions

Abstraction

DynOptInterface.AbstractBoundaryFunction — Type

AbstractBoundaryFunction

Abstract super-type for dynamic functions evaluated at phase boundaries.

That is, expressions that may contain:

$t_i^0$ Initial{PhaseIndex}

$t_i^f$ Final{PhaseIndex}

$y_j(t^0)$ Initial{DynamicVariableIndex}

$y_j(t^f)$ Final{DynamicVariableIndex}

Initial & Final

DynOptInterface.Initial — Type

Initial{DF}(evaluand::DF) where {DF<:AbstractDynamicFunction}

Represents the evaluation of an AbstractDynamicFunction at the initial point of its phase.

It is a sub-type of AbstractBoundaryFunction. The dynamic function is stored in the evaluand field.

DynOptInterface.Final — Type

Final{DF}(evaluand::DF) where {DF<:AbstractDynamicFunction}

Represents the evaluation of an AbstractDynamicFunction at the final point of its phase.

It is a sub-type of AbstractBoundaryFunction. The dynamic function is stored in the evaluand field.

Linkage

DynOptInterface.Linkage — Type

Linkage{DF}(
    final_evaluand::DF,
    initial_evaluand::DF,
) where {DF<:AbstractDynamicFunction}

Represents the expression $f_f(y(t^f), t^f) - f_0(y(t^0), t^0)$.

It is a sub-type of AbstractBoundaryFunction. The final function is stored in the final_evaluand field and the initial function is stored in the initial_evaluand field.

Nonlinear

DynOptInterface.NonlinearBoundaryFunction — Type

NonlinearBoundaryFunction(head::Symbol, args::Vector{Any})

Represents a general function $f_b(y(t^0), y(t^f), t^0, t^f, x)$.

It is a sub-type of AbstractBoundaryFunction. Similar to MOI.ScalarNonlinearFunction, this function is represented by an expression tree, using the following fields:

head

The symbol head must be an operator that is supported by the model. The model attribute MOI.ListOfSupportedNonlinearOperators provides a list of supported operators. If the optimizer does not support head, an MOI.UnsupportedNonlinearOperator error is thrown.

args

The vector args contains the arguments to the nonlinear operator. The possible arguments that may be included are:

A constant value of type T<:Real

An MOI.VariableIndex $x_k$

An MOI.ScalarAffineFunction $a^\top x + b$

An MOI.ScalarQuadraticFunction $x^\top Q x + a^\top x + b$

An MOI.ScalarNonlinearFunction $f(x)$

An Initial{PhaseIndex} $t_i^0$

A Final{PhaseIndex} $t_i^f$

An Initial{DynamicVariableIndex} $y_j(t_i^0)$

A Final{DynamicVariableIndex} $y_j(t_i^f)$

Another NonlinearBoundaryFunction

Additionally, the optimizer must indicate support of argument types through the supports_objective_argument and supports_constraint_argument functions. Otherwise UnsupportedObjectiveArgument and UnsupportedConstraintArgument errors are thrown.

Integrals

DynOptInterface.Integral — Type

Integral{DF}(integrand::DF) where {DF<:AbstractDynamicFunction}

Represents the integral $\int_{t_i^o}^{t_i^f} f_d(\dot{y}(t_i), y(t_i), t_i, x) \mathrm{d}t_i$.

It is a sub-type of AbstractBoundaryFunction. The integrand is stored in the integrand field.

DynOptInterface.MultiPhaseIntegral — Type

MultiPhaseIntegral{DF}(
    integrands::Vector{DF},
) where {DF<:AbstractDynamicFunction}

Represents the sum of integrals $\sum_i \big[ \int_{t_i^o}^{t_i^f} f_d(\dot{y}(t_i), y(t_i), t_i, x) \mathrm{d}t_i \big]$.

It is a sub-type of AbstractBoundaryFunction. The vector of integrands is stored in the integrands field.

DynOptInterface.Bolza — Type

Bolza{BF,IF}(
    bou_fun::BF,
    integral::IF,
) where {BF<:AbstractBoundaryFunction,IF<:Union{Integral,MultiPhaseIntegral}}

\[f_b(y_0, y_f, t_0, t_f, x) + \int_{t_i^o}^{t_i^f} f_d(\dot{y}(t_i), y(t_i), t_i, x) \mathrm{d}t_i\]

Represents the sum of an AbstractBoundaryFunction with either an Integral or a MultiPhaseIntegral.

It is a sub-type of AbstractBoundaryFunction. The boundary function is stored in the bou_fun field and the integral is stored in the integral field.