Dynamic Variables

DynOptInterface.DynamicVariableIndex — Type

DynamicVariableIndex(value::Int64, phase::PhaseIndex)

Represent the dynamic variable $\boldsymbol{y}_j(\cdot)$ in an expression.

It is a subtype of AbstractDynamicFunction. To allow for deletion, indices need not be consecutive.

Dynamic Variables in Models

DynOptInterface.supports_dynamic_variables — Function

supports_dynamic_variables(model::MOI.ModelLike)::Bool

Indicate whether model supports dynamic variables.

DynOptInterface.UnsupportedDynamicVariablesError — Type

UnsupportedDynamicVariablesError(message::String)

The model does not support dynamic variables.

That is, supports_dynamic_variables returns false.

DynOptInterface.add_dynamic_variable — Function

add_dynamic_variable(model::MOI.ModelLike, phase::PhaseIndex)::DynamicVariableIndex

Add a dynamic variable to model, returning a DynamicVariableIndex.

An AddDynamicVariableNotAllowedError is thrown if a dynamic variable cannot be added to model in its current state.

DynOptInterface.AddDynamicVariableNotAllowedError — Type

AddDynamicVariableNotAllowedError(message::String)

Dynamic variables cannot be added to the model in its current state.

MathOptInterface.is_valid — Method

MOI.is_valid(model::MOI.ModelLike, dyn_var::DynamicVariableIndex)::Bool

Indicate whether dyn_var refers to a valid DynamicVariableIndex in model.

DynOptInterface.InvalidDynamicVariableError — Type

InvalidDynamicVariableError(dyn_var::DynamicVariableIndex)

The dynamic variable dyn_var is not valid in the model.

Dynamic Variable Attributes

DynOptInterface.AbstractDynamicVariableAttribute — Type

AbstractDynamicVariableAttribute

Supertype for dynamic variable attributes.

MathOptInterface.supports — Method

MOI.supports(
    model::MOI.ModelLike,
    attr::AbstractDynamicVariableAttribute,
)::Bool

Indicate whether model supports the dynamic variable attribute attr.

MathOptInterface.set — Method

MOI.set(
    model::MOI.ModelLike,
    attr::AbstractDynamicVariableAttribute,
    dyn_var::DynamicVariableIndex,
    value::Any,
)

Assign value to the attribute attr of dynamic variable dyn_var in model model.

MathOptInterface.get — Method

MOI.get(
    model::MOI.ModelLike,
    attr::AbstractDynamicVariableAttribute,
    dyn_var::DynamicVariableIndex,
)

Return the value of the attribute attr set to dynamic variable dyn_var in model model.

If the attribute attr is not supported by model then an error should be thrown. If the attribute is supported but has not been set, nothing is returned.

DynOptInterface.DynamicVariableName — Type

DynamicVariableName

A dynamic variable attribute for a String identifying a dynamic variable.

DynOptInterface.DynamicVariableInitialStart — Type

DynamicVariableInitialStart

A dynamic variable attribute for the start value of the dynamic variable at the initial phase boundary.

DynOptInterface.DynamicVariableFinalStart — Type

DynamicVariableFinalStart

A dynamic variable attribute for the start value of the dynamic variable at the final phase boundary.