Phases

DynOptInterface.PhaseIndex — Type

PhaseIndex(value::Int64)

A type-safe wrapper for Int64 for use in referencing phases in a model.

It is a sub-type of AbstractDynamicFunction. It represents an independent variable $t_i$ in a domain $[t_i^0, t_i^f]$. The Int64 index value is stored in the value field. To allow for deletion, indices need not be consecutive.

DynOptInterface.supports_phase — Function

supports_phase(model::MOI.ModelLike)::Bool

Returns a Bool indicating whether model supports phases.

DynOptInterface.UnsupportedPhase — Type

UnsupportedPhase(message::String)

An error indicating that phases are not supported by the model.

That is, that supports_phase returns false. The String error message is stored in the message field.

DynOptInterface.AddPhaseNotAllowed — Type

AddPhaseNotAllowed(message::String)

An error indicating that phases cannot be added to the model in its current state.

The String error message is stored in the message field.

DynOptInterface.add_phase — Function

add_phase(model::MOI.ModelLike)::PhaseIndex

Add a phase to model, returning a PhaseIndex.

An AddPhaseNotAllowed error is thrown if a phase cannot be added to the model in its current state.

MathOptInterface.is_valid — Method

MOI.is_valid(model::MOI.ModelLike, index::PhaseIndex)::Bool

Return a Bool indicating whether index refers to a valid PhaseIndex in model.

DynOptInterface.InvalidPhaseIndex — Type

InvalidPhaseIndex(index::PhaseIndex)

An error indicating that the phase index is not valid.

Attributes

The PhaseIndex object is compatible with the following attributes:

MOI.VariableName with value types String

MOI.VariablePrimal with value types Tuple{<:Real,<:Real}

MOI.VariablePrimalStart with value types Tuple{<:Real,<:Real}

MathOptInterface.supports — Method

MOI.supports(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    ::Type{PhaseIndex},
)::Bool

Return a Bool indicating whether model supports the attribute attr for PhaseIndexs.

MathOptInterface.set — Method

MOI.set(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    index::PhaseIndex,
    value,
)

Assign value to the attribute attr of phase index in model model.

An MOI.UnsupportedAttribute error is thrown if model does not support the attribute attr, and a MOI.SetAttributeNotAllowed error is thrown if it supports the attribute attr but it cannot be set.

MathOptInterface.get — Method

MOI.get(
    model::MOI.ModelLike,
    attr::MOI.AbstractVariableAttribute,
    index::PhaseIndex,
)

Return the value of the attribute attr set to phase index in model model.

If the attribute attr is not supported by model then an error should be thrown. If the attribute is supported but has not been set, nothing is returned.