Phases

DynOptInterface.AbstractDynamicFunction — Type

AbstractDynamicFunction

Supertype for scalar-valued dynamic functions.

That is, expressions that contain a phase parameter $t^{(i)}$. Each dynamic function must be defined on one (and only one) phase.

DynOptInterface.PhaseIndex — Type

PhaseIndex(value::Int64)

Represent the phase parameter $t^{(i)}$ in an expression.

It is a subtype of AbstractDynamicFunction. To allow for deletion, index values need not be consecutive.

DynOptInterface.phase_index — Function

phase_index(dyn_fun::AbstractDynamicFunction)::PhaseIndex

Return the PhaseIndex $t^{(i)}$ of a dynamic function dyn_fun.

DynOptInterface.NonUniquePhaseError — Type

NonUniquePhaseError(message::String)

A dynamic function was not defined on one (and only one) phase.

Phases in Models

DynOptInterface.supports_phases — Function

supports_phases(model::MOI.ModelLike)::Bool

Indicate whether model supports phases.

DynOptInterface.UnsupportedPhasesError — Type

UnsupportedPhasesError(message::String)

The model does not support phases, that is, supports_phases returns false.

DynOptInterface.add_phase — Function

add_phase(model::MOI.ModelLike)::PhaseIndex

Add a phase to model, returning a PhaseIndex.

An AddPhaseNotAllowedError is thrown if a phase cannot be added to the model in its current state.

DynOptInterface.AddPhaseNotAllowedError — Type

AddPhaseNotAllowedError(message::String)

Phases cannot be added to the model in its current state.

MathOptInterface.is_valid — Method

MOI.is_valid(model::MOI.ModelLike, phase::PhaseIndex)::Bool

Indicate whether phase refers to a valid PhaseIndex in model.

DynOptInterface.InvalidPhaseError — Type

InvalidPhaseError(phase::PhaseIndex)

The phase phase is not valid in the model.

Phase Attributes

DynOptInterface.AbstractPhaseAttribute — Type

AbstractPhaseAttribute

Supertype for phase attributes.

MathOptInterface.supports — Method

MOI.supports(
    model::MOI.ModelLike,
    attr::AbstractPhaseAttribute,
)::Bool

Indicate whether model supports the phase attribute attr.

MathOptInterface.set — Method

MOI.set(
    model::MOI.ModelLike,
    attr::AbstractPhaseAttribute,
    phase::PhaseIndex,
    value::Any,
)

Assign value to the attribute attr of phase phase in model model.

MathOptInterface.get — Method

MOI.get(
    model::MOI.ModelLike,
    attr::AbstractPhaseAttribute,
    phase::PhaseIndex,
)

Return the value of the attribute attr set to phase phase in model.

If the attribute attr is not supported by model then an error should be thrown. If the attribute is supported but has not been set, nothing is returned.

DynOptInterface.PhaseName — Type

PhaseName

A phase attribute for a String identifying a phase.

DynOptInterface.PhaseInitialStart — Type

PhaseInitialStart

A phase attribute for the start value of the initial phase boundary.

DynOptInterface.PhaseFinalStart — Type

PhaseFinalStart

A phase attribute for the start value of the final phase boundary.